Added specific indexing for shear and moment array
The shear and moment in the second span were outputting values for the
first span locations. The values for shear and moment were junk values.
Emma wrote an integrated solution that used the analysis node value to
come up with the correct array index to store the shear or moment value.

This funcationality was buried in the shear or moment envelope function. I
wanted something that was more transparent. Emma's solution gave me
insight into de-coupling the analysis node value and the span index array
value.

This solution is the work in progress from the "Two span output incorrect
- In progress" commits.