Reorganized code, single mlob file was to long
Broke the single mlob.py file into four more manageable files that make
sense from a program structure stand point.

The main.py file manages the execution of the other three files.

The user_input.py file is the user interface. This will probably grow
in the future and needs its own place to live. It had to be named
"user_input" because "input" is a built-in fuction.

The mlob.py file is the main number crunching analysis engine. It modifies
the user input for ease of analysis, adds the trailing load to the main
train, creates analysis nodes, calculates the shear, moment, and pier
reaction for each node along the span, and envelopes these forces.

The output.py file is the formatted output for the user to use. It echoes
the input and provides a pretty print of the enveloped forces.