Moved test_short to test_mlob. Deleted test_long.
Learned more about unit testing working on another project and decided
that I don't need multiple testing files at this stage. I can write all
the testing in one file and then use testing suites to break it out into
shorter testing and longer testing.

The shorter testing will be on the "helper" functions and maybe one test
on the analyze_vehicle function for a shorter span because it can take
a significant amount of time to run for longer spans.

The longer testing will be a more in-depth test of the analyze_vehicle
function. The plan is to check every span length given in AREMA Section
1.15.1 which will take a long time to calculate.