Created new function for span begin and end coordinates.
Functionality was at the top of the analysze_vehicle function. Realized it
needed to be moved when I started writing the test for the node_location
function.

The span_begin_end_function needs to include some exceptions to
make sure there are no negative lengths provided by the user. This would
really mess up subsequent functions that use those values and it didn't
make sense to add the exceptions in the analyze_vehicle function.