Added an optional parameter to add_trailing_load function.
Decided that at some point in the future it may be advantageous to access
the spacing used to approximate the trailing distributed load as discrete
point loads. Added an optional parameter that defaults to 0.5. This is
reasonable as long as the user is entering in feet, which should typically
be the case.