Add progress bars to the mlob module with tqdm.
Some runs can take a 60 seconds or more to complete. The user may think
there is something wrong and interrupt the program before it can complete
the calculations. The tqdm is a very low overhead progress bar.

I added one to the main loop but it only has two iterations so it jumps to
50% complete then 100%, this is the direction switching loop. I added another
one to the next loop down it was still jumping to comletion points, this
is the analysis node loop. I added one more to next loop down and now it
shows a constant working progress bar, this is the axle stepping loop. I
think anything lower than this is probably not useful output for the user.