Removed unit labels from output tables. Specified span for max reaction.
The program is independent of units and it is up to the user to keep track of
the units and be consisent. Therefore, I removed the unit labels from the output
tables.

I noticed that the Vmax and Mmax output doesn't make sense without knowing which
span it corresponds to and it is not obvious which span it does correspond to.
Clarified which span each maximum corresponds to.