Added documentation. Clarified modifications to axle_spacing and axle_wt.
Added documentation for calc_load_and_loc and move_axle_loc.

The modifications done in add_trailing_loads to the axle_spacing and axle_wt
lists is not obvious. Returned them from the function instead of implicitly
over-writing their values to make it more obvious that these list are updated in
that function.