Added basic testing functions for calc_pier_reaction.
Finished testing basic functionality of the calc_pier_reaction function.
Used excel functionality to independently check the output values. Saved a
copy of the excel files in the test folder.

Added an additional testing function for the calc_load_and_loc to test the
"rtl" functionality. The "ltr" was already tested.

All tests are successful.