Added testing for calc_shear. Started other testing for span 2.
Added two testing function for calc_shear. Both in span 1, one for ltr and one
for rtl directions. The test was checked by using excel functionality to
calculate the value of the shear.

Removed the "Re" parameter from the calc_shear function. It is not used in the
function and therefore not needed. It could be confusing to someone else reading
the code or me later on.

calc_reactions.xlsx was not deleted just moved to the testing folder. I thought
I already commited this change but apparently not.

All other excel file changes were due to modifying the input to calculate the
shear.

All tests passed.