Added tests for calc_reactions and calc_pier_reactions in span 2.
Added code for multiple new tests in span 2. Completed tests for
calc_reactions and calc_pier reactions with x=125 so it is testing the
functionality in span 2.

Four other tests have not been completed in span 2. calc_shear with direction ltr,
and calc_reactions, calc_pier_reactions, and calc_shear with direction rtl. The
basic code structure is there but the independent calculations have not been
completed to produce an independent answer for testing purposes.

Four tests fail.