Testing for calc_moment function complete.
Finished tests for calculating the moment in each span.

Created new excel file manual_mlob.xlsx. This is a simple combination of each of the
"functions" in excel put together for easier testing. Should have done this a
long time ago.

All tests pass.

Still trying to figure out how to test the envelope functions and the
analyze_vehicle function.

Need to start writing the longer test to replicate the AREMA table.