Added documentation to calc_shear, calc_moment, and analyze_vehicle.
Realized there wasn't documentation for these functions. Wrote documentation for
calc_shear and calc_moment.

Discussed with Emma the documentation for analyze_vehicle. We decided that it
should be documented based on what it is returning despite the fact that it is
using every function in the module to accomplish this. Wrote out in English what
the analyze_vehicle function is doing and why a minimum number of nodes is
required for anlaysis.