Discovered issues while testing mlob against AREMA table.
First, python passes lists by reference not by value so the axle_spacing and
axle_wt lists were being manipulated on each subsequent run and the results were
incorrect.

Fixed by creating a copy of the axle_spacing and axle_wt as follows:
mod_axle_spacing = axle_spacing[:]
mod_axle_wt = axle_wt[:]

Also updated the variable names because it is confusing to have the same
variable name "axle_spacing" or "axle_wt" when the variable is has been changed.

Second, it was discovered that the pier/floorbeam reaction is not being
calculated correctly. This is the problem with checking your own work with
another solution you developed. It is easy to make the same mistake twice.

I think there will have to a calc_load_and_loc function for the span moments,
shear, and reactions, and a similar calc_load_and_loc for the pier reaction.