Fixed pier reaction problem.
Wrote new function that omits loads on the ends of the span. The shear and
moment on the span can be calculated based on a total load on the span. However,
this does not work for the reaction at the center pier for two adjacent spans.

Also had to move the Rpier initialization variable further up in the nested
for-loops. It was being re-initialized for each node instead of persisting for
the entire span.

20 of the basic test don't work right now. Will be fixing this shortly.