Fixed problem with shear calculation, added line to output file.
The shear in span 2 was calculating improperly in some instances. The reason for
this was the improper rounding of the node location coordinates. This was fixed
by rounding the node locations to two decimal places. More testing is required
but it appears based on preliminary testing that this works.

I noticed the output was not properly described. The loads output are for one
rail or 1/2 track or one line of wheels or 1/2 an axle. This is the standard for
the AREMA Tb 15-1-16 and I want the output to be clear.

Forgot to run the basic tests. Will run after commit to see if they still pass.