Added user_test.py for quick user verification.
Allows the user to quickly run the program for the E80 and Alternate Live Loads
from AREMA and compare that directly to the values from the AREMA Table. It
also provides an error with respect to the AREMA Table.

This has not been run and debugged.

This is supposed to mimic the test_mlob_long except it is supposed to be used by
the end user to quickly verify the program.