On-going problem with unsymmetric shear.
If the calc_shear function in mlob has an absolute value for both the ltr and
rtl the spikes for the span 1 go away. The spikes for span 2 are still present.

Thinking on this I am not sure if this is the correct solution.

I am not real sure what the underlying problem is and I am still trying to
figure it out.

Made some edits to manual_mlob hoping it would help reveal the error but to no
avail yet.

Added a graphing function to dev_test so that it is easier to see the symmetry
and the spikes for each span. I really like pyplot, it's so fast to make really
good looking plots.