I think the problem with the shear is a numerical rounding problem.
For some reason the analysis node location or some other location is being round
in a weird way and causing the reaction or total load on the span (I can't
remember which) to tank. This is screwing up the value for the shear.

At least I have an idea about what the problem is now.

Still working on a fix.