Solved problem with spikes in the shear envelope.
I suspected that it was a problem with decimal fractions not be precisely
represented and subsequent comparison of values (I did a poor job of describing
this in my previous commit). Lo-and-behold, the axle location was off by
something like 0.00000001 so it was technically greater than the node location
and it botched the calculation of the shear at that location.

I am still not sure why this didn't effect the moment calculation at that point.
Maybe it did and I didn't notice it.

Anywhos, I fixed it by rounding the calculated axle locations to 3 decimal
places. There was some something similar done for the analysis node locations as
well with a previous problem (I am sure if you look through the log there would
be mention of it).

Fixed.

All tests pass.