Added the start of some graphing capabilities. Far from complete.
In figuring out how to fix the spikes in the shear envelope diagram I added a
small function the dev_test file to graph span 1 and span 2 shear envelopes. I
was tired of copying and pasting data to excel to graph it.

I had already added an enhancement issue to the github project for v1.2 (that's
out in the future) to add graphing capabilities. This is the start of that. I
saw the code and an opportunity to quickly capitalize and I did. Instead of not
looking at this for sometime and completely forgetting about it.

It would probably work as is with some formatting to the graphs and adding a way
for the user to request the generation of the graphs.

Saving it for another day.