Using for actual design calculations.
Needed to run 170ft and 100ft span to get the pier reaction. Checked max
moments, shear, and pier reactions with another program. All in good agreement.

Really need to implement the quick input of the standard CE80 and Alternate
loads. It would be a real PIA to input those each time a user wants to make a
run since those are by far the most common axle configurations.

I would also be nice to allow the user to output a single enveloped output or
the independent output for both configurations. The analysis is there, just have
to make use of it in a good way.