Updated corresponding shear calculation. Having rounding error issues.
For some reason the rounding that is being done to the node locations is off
just slightly and it is making the corresponding shear in this specific case
unsymmetric because the node is +/- just enough to round up or down.

It's almost like it needs some kind of tolerance built in to look at +0.001 and
-0.001 around the node location. This is crazy.

There has to be an easier way.