Added functionality to output the controlling axle location.
Added new functionality to track the controlling axle locations for maximum
shear and moment at each node, and the maximum pier reaction if there are two
spans. Currently only the axle location for the maximum shear, moment, and
pier reaction (if applicable) for each span are written in the output file.

The direction of the train is also included in this information because the
train and the spans may not be symmetric.

I have propogated this through all testing.

All tests pass.

Although I have not seen the output with my own two eyes for a two span
structure. So it is not 100% tested but it appears work (famous last words).

Still working on the corresponding shear.