I think I finally fixed the corresponding forces problem.
I noticed earlier today that the shears and moments were off in the user_test
and the mlob_test_long wouln't run. Something was buggy.

First I had to fix the problem of the last value node location value in the
span. All of the incremental errors add up to get to this last value and
sometimes it doesn't quite equal the actual value at the end of the span. I
thought I fixed this by adding a little bit of logic to the node_locaiton
function. The problem was I used the actual span length instead of the location
of the node in the global coordinate system. This will break everything for a 2
span which is what was happening with the tests. However, the dev_test I was
running was for two spans. Maybe I should run two spans more often....

The second problem was the rounding of the shear and moment calculations. The
alorithm is such that it updates the value of the corresponding shear or moment
with the maximum value of the maximum moment or shear, respectively. The problem
arises when the value gets rounded up or down slightly throwing off the maximum
value and then updating the corresponding value incorrectly. I noticed this when
looking at the maximum moment values at the end of the span sometimes a value of
10^-12 would be generated which is greater than zero and incorrectly the
corresponding shear to this rounding error number would be inserted. To correct
for this I decided to round all the shear and moment calculations. It seems to
have fixed the problem.

All tests pass.