Changing folder and file naming convention.
The repo on Github is now call 'mlob' and the main folder for code is
now also called 'mlob'. It seems to fall more in line with other
python Github repos.

Also changing the name of the mlob.py file to analysis.py. I think it is
more descriptive about what that file does.