Delete reporting R1, R2, axle num for pier reaction
Realized that the values being reported for the  the beginning and
end reactions for a two span structure were not valid. They are
used for calculating the center pier reaction but don't apply to
the reactions at the ends.

Also removed reporting the axle position for the pier reaction.

Want to get it working again, may add again later.