Fixed! Shear works, moment works.
Hopefully for the last time; until I start adding *features*.

So apparently what had to be done was calculate the shear on both
side of the analysis node. Which makes sense once you think about
it. The point load is a discontinuity in the shear force. So, to
accurately predict the maximum shear force you have to look at
the discontinuity approaching from each side.

To do this in the code you have to expand the
calc_load_and_location function to *include* and *exclude* the
point load at the analysis node. Effectively, calculating the
shear just before and just after the point load. Or another way
to think about is that there are two shear values at a point
load one *before* and one *after*, and you need to *envelope*
those values.

Guess I never thought about it that hard. Drawing a shear
diagram by hand it doesn't explicitly come up. And I don't
think I have ever thought about the value of shear in a
structure exactly at a conctrated load. Shear before and after
but never exactly at one and how you would need to consider
the envelope of the values, or the values just before and
just after.

This is a reason why I enjoy structural engineering. You think
you really understand a concept but there is always something
more to learn and the power in the simplicity of basic concepts.
Just fascinating.